+++ frontend/src/App.js
@@ -2,8 +2,8 @@ import React from "react";
 import { Switch, Route } from "react-router-dom";
 import "./App.css";
 import Identification from "./pages/Identification";
-import Login from "./components/Login";
-import Signup from "./components/Signup";
+import Login from "./pages/Login";
+import Signup from "./pages/Signup";
 import HomePage from "./components/HomePage";
 import NavBar from "./components/NavBar";
