+++ frontend/src/pages/Login.jsx
@@ -1,11 +1,10 @@
 import React, { useState } from "react";
 import "./style/LoginSignup.scss";
-import IdentificationHeader from "./IdentificationHeader";
+import IdentificationHeader from "../components/IdentificationHeader";
 
-export default function Signup() {
+export default function Login() {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
   return (
     <div id="loginSignup">
       <IdentificationHeader />
@@ -28,16 +27,6 @@ export default function Signup() {
             type="password"
             value={password}
             onChange={() => setPassword()}
-          ></input>
-        </label>
-        <label className="button">
-          <input
-            id="confirmPassword"
-            name="confirmPassword"
-            placeholder="mot de passe"
-            type="password"
-            value={confirmPassword}
-            onChange={() => setConfirmPassword()}
           ></input>
         </label>
         <input className="button" type="submit" value="Soummettre"></input>
