+++ frontend/src/pages/Signup.jsx
@@ -1,10 +1,11 @@
 import React, { useState } from "react";
 import "./style/LoginSignup.scss";
-import IdentificationHeader from "./IdentificationHeader";
+import IdentificationHeader from "../components/IdentificationHeader";
 
-export default function Login() {
+export default function Signup() {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+  const [confirmPassword, setConfirmPassword] = useState("");
   return (
     <div id="loginSignup">
       <IdentificationHeader />
@@ -27,6 +28,16 @@ export default function Login() {
             type="password"
             value={password}
             onChange={() => setPassword()}
+          ></input>
+        </label>
+        <label className="button">
+          <input
+            id="confirmPassword"
+            name="confirmPassword"
+            placeholder="mot de passe"
+            type="password"
+            value={confirmPassword}
+            onChange={() => setConfirmPassword()}
           ></input>
         </label>
         <input className="button" type="submit" value="Soummettre"></input>
